Gestion erreurs 403 forbidden pour articles & recettes côté front & back + retouche interceptor pour erreurs

+++ apps/back/src/app/articles/articles.service.ts
@@ -1,20 +1,30 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { ItemBase, ItemType } from '@shared-interfaces/items';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ArticlesEntity } from './articles.entity';
 import { Repository } from 'typeorm';
 import { ArticlesCreateDto, ArticlesUpdateDto } from './articles.dto';
+import { DiversService } from '../calendar/divers/divers.service';
+import { DaysService } from '../calendar/days/days.service';
+import { CoursesService } from '../courses/courses.service';
+import { RecettesService } from '../recettes/recettes.service';
 
 @Injectable()
 export class ArticlesService {
-  constructor(@InjectRepository(ArticlesEntity) private _articlesEntityRepository: Repository<ArticlesEntity>) {}
+  constructor(
+    @InjectRepository(ArticlesEntity) private _articlesEntityRepository: Repository<ArticlesEntity>,
+    private _diversService: DiversService,
+    private _daysService: DaysService,
+    private _coursesService: CoursesService,
+    private _recettesService: RecettesService,
+  ) {}
 
   async getArticles(): Promise<ItemBase[]> {
     const query = this._articlesEntityRepository
       .createQueryBuilder('a')
       .select('*')
       .where('a.id is not null')
-      .orderBy({ id: 'ASC'})
+      .orderBy({ id: 'ASC' })
       .getRawMany();
 
     if (!query) {
@@ -40,8 +50,24 @@ export class ArticlesService {
     return this._articlesEntityRepository.findOneBy({ id });
   }
 
- async removeArticle(id: number): Promise<void> {
-   const entity = await this._articlesEntityRepository.findOneBy({id});
-   await this._articlesEntityRepository.remove(entity);
- }
+  async removeArticle(id: number): Promise<void> {
+    const entity = await this._articlesEntityRepository.findOneBy({ id });
+    if (!entity) {
+      throw new NotFoundException();
+    }
+
+    const existInDivers = await this._diversService.removeForbiddenIfExisting(id, ItemType.ARTICLE);
+    const existInDays = await this._daysService.removeForbiddenIfExisting(id, ItemType.ARTICLE);
+    const existInCourses = await this._coursesService.removeForbiddenIfExisting(id);
+    const existInRecettes = undefined;
+    // TODO faire les recettes quand relations OK ! :)
+    // const existInRecettes = await this._recettesService.removeForbiddenIfExisting(id);
+
+    if (!!existInDivers || !!existInDays || !!existInCourses || !!existInRecettes) {
+      const existIn = [existInDivers, existInDays, existInCourses, existInRecettes].filter(item => !!item);
+      throw new ForbiddenException(`Article utilisé dans ${existIn.join(', ')}`);
+    }
+
+    await this._articlesEntityRepository.remove(entity);
+  }
 }
