+++ apps/back/src/app/recettes/recettes.service.ts
@@ -1,13 +1,19 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { ArticleTags, ItemBase, ItemType, RecetteTags } from '@shared-interfaces/items';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import { ItemBase, ItemType } from '@shared-interfaces/items';
 import { RecettesCreateDto, RecettesUpdateDto } from './recettes.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { RecettesEntity } from './recettes.entity';
+import { DiversService } from '../calendar/divers/divers.service';
+import { DaysService } from '../calendar/days/days.service';
 
 @Injectable()
 export class RecettesService {
-  constructor(@InjectRepository(RecettesEntity) private _recettesEntityRepository: Repository<RecettesEntity>) {}
+  constructor(
+    @InjectRepository(RecettesEntity) private _recettesEntityRepository: Repository<RecettesEntity>,
+    private _diversService: DiversService,
+    private _daysService: DaysService,
+  ) {}
 
   async getRecettes(): Promise<ItemBase[]> {
     const query = this._recettesEntityRepository
@@ -42,6 +48,26 @@ export class RecettesService {
 
   async removeRecette(id: number): Promise<void> {
     const entity = await this._recettesEntityRepository.findOneBy({ id });
+    if (!entity) {
+      throw new NotFoundException();
+    }
+
+    const existInDivers = await this._diversService.removeForbiddenIfExisting(id, ItemType.RECETTE);
+    const existInDays = await this._daysService.removeForbiddenIfExisting(id, ItemType.RECETTE);
+
+    if (!!existInDivers || !!existInDays) {
+      const existIn = [existInDivers, existInDays].filter(item => !!item);
+      throw new ForbiddenException(`Recette utilisée dans ${existIn.join(', ')}`);
+    }
+
     await this._recettesEntityRepository.remove(entity);
+  }
+
+  async removeForbiddenIfExisting(id: number): Promise<string> {
+    const existing = await this._recettesEntityRepository.findOneBy({ id: id });
+    // TODO indiquer quelle recette
+    if (existing) {
+      return 'recettes';
+    }
   }
 }
